Split bot functionality into multiple modules
The corresponding change to tests remains to be done.
