+++ run-scheduled.py
@@ -11,8 +11,10 @@ from conf.header import HEADER
 from conf import conf
 from conf.secrets import habitica_credentials
 from habot.birthdays import BirthdayReminder
-from habot.bot import (handle_PMs, SendWinnerMessage, AwardWinner,
-                       CreateNextSharingWeekend, UpdatePartyDescription)
+from habot.functionality.react import handle_PMs
+from habot.functionality.sharing_weekend_challenge import (
+        SendWinnerMessage, CreateNextSharingWeekend, AwardWinner)
+from habot.functionality.party_description import UpdatePartyDescription
 from habot.exceptions import CommunicationFailedException
 from habot.io.messages import HabiticaMessager, PrivateMessage
 from habot.habitica_operations import HabiticaOperator
